Do not remove fifo directory if keep sandbox is True

+++ cms/grading/tasktypes/Communication.py
@@ -3,7 +3,7 @@
 
 # Contest Management System - http://cms-dev.github.io/
 # Copyright © 2010-2012 Giovanni Mascellani <mascellani@poisson.phc.unipi.it>
-# Copyright © 2010-2012 Stefano Maggiolo <s.maggiolo@gmail.com>
+# Copyright © 2010-2014 Stefano Maggiolo <s.maggiolo@gmail.com>
 # Copyright © 2010-2012 Matteo Boscariol <boscarim@hotmail.com>
 # Copyright © 2012-2014 Luca Wehrstedt <luca.wehrstedt@gmail.com>
 #
@@ -264,4 +264,5 @@ class Communication(TaskType):
 
         delete_sandbox(sandbox_mgr)
         delete_sandbox(sandbox_user)
-        rmtree(fifo_dir)
+        if not config.keep_sandbox:
+            rmtree(fifo_dir)
